CSG: new file 'csg_nukem' for the DUKE NUKEM 3D specific code.
I have decided to separate the DOOM and DUKE3D CSG code, due to a
large difference in the way linedefs (aka walls) are handled.  In
DOOM they Can be treated fairly independently (linedefs are the
top of the dependency tree), whereas DUKE3D needs to maintain wall
loops and the focus is more on sectors.