Build: for '?' substitution, support four simple operators: '+', '-', '==' and '!='.  They all require a numeric operand, no whitespace is allowed.
The equality and inequality operators produce 1 for true, 0 for false.
Their main purpose will be for conditionally present/absent stuff in
prefabs.

Note: more operators or more complex expressions is NOT planned, as this
can be handled by using function values, although so far this feature has
not been needed at all and I'm hoping it'll stay that way :-)