Build: implemented conditional prefab elements via 'only_if' field.
When the 'only_if' field has a zero, false or empty-string value then
the particular brush, model or entity will not be rendered (it gets
removed from the prefab copy).

For brushes, the 'only_if' field must occur in an 'm' table at the
beginning of the brush -- it does NOT work in coordinate tables.