Hallways: created 'HALL_BASIC_I_LIFT' prefab and a skin for DOOM.
This was complicated since I wanted the lift to be raise-wait-lower
but there is no such linedef special in DOOM (or even BOOM).  The
solution was to have a line in the middle of the lift to raise up
the floor (W1 for once only), and the lines around the lift itself 
have the normal lift specials.