C++ / Lua stuff: changed Script_Open() function to create the lua_State and then load everything for a particular game into it (including the scripts/, modules/ and engines/ files).  The Script_Close() function closes that lua_State and frees everything.
Eventually this will be used to only load the definitions for the
currently chosen game (unlike before where we loaded all game defs).

Also we change the package path to be "xxx/scripts/?.lua" when loading
the normal scripts, and "xxx/<game_dir>/?.lua" when loading the
game-specific scripts.  That will allow using the same filename in
both contexts (e.g. "monsters.lua" can be used in both).