Merged in v3.57 level layout logic, replacing the v5.10 prefab layouter.
This work was done in a separate repository, adjusting the old code to
function correctly in the new environment.  There are still much to do
before the merger is really finished, especially handling of lighting,
and numerous features from V4/V5 need to be ported to this code-base
(which could use a lot of tidying up too).