Monsters : new "weap_min_damage" field replaces "min_weapon".
Changed this because the "min_weapon" field was unreliable, it
used the "level" field of weapons as an indicator of strength
when it only indicates how soon to introduce the weapon.

This lead to a bug (topic 540 on the forum) which caused Boss
monsters to be lacking from levels when using the Weapon Control
module and setting most weapons to a custom setting.