PhysFS : fixed ZIP_seek due to a breaking change in zlib library.
The zlib library introduced a check that a z_stream object is at
the same address as when it was initialized.  Code in ZIP_seek was
creating a new z_stream object and then memcpy'ing it over the old
z_stream object, which caused future ZIP_read calls to fail due to
inflateStateCheck() in the zlib code.

This commit makes the rewind logic in ZIP_seek simpler, and just
recreates the existing z_stream object to do the rewind.  This
has fixed the problem of reading wad files in Oblige addons when
using zlib library version is newer than 1.2.8.