Various prefabs fixes
Allowed monster to travel the elevator shafts in both directions. Fixed a point switch that could get the player stuck if he could not jump, reduced the amount of torches on 2 door prefabs (fps optimization for expensive mods), pushed the gameshow doors further into the fab to make it stop showing on the other side of a thin wall, forced monsters out of the item elevator shafts reducing the chances of seeing monsters from top down.